requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==2.0.12
Django==3.2
django-filter==23.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
flake8-isort==6.1.1
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pluggy==0.13.1
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.26.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
